Change Unified::field to return Vec<f64>
Betaflight stores integers, PX4/ArduPilot store floats. Vec<i64>
truncated all sub-1 deg/s gyro values to zero. Vec<f64> preserves
full precision across all formats — BF integers promote losslessly
(all values < 2^53). No unit scaling hacks needed.

Also adds vehicle_angular_velocity as PX4 gyro fallback topic
and fixes PX4 motor_count to check non-zero outputs.

+++ propwash-core/src/types.rs
@@ -292,12 +292,12 @@ pub trait Unified {
     fn craft_name(&self) -> &str;
     fn sample_rate_hz(&self) -> f64;
     fn duration_seconds(&self) -> f64;
-    fn field(&self, field: &SensorField) -> Vec<i64>;
+    fn field(&self, field: &SensorField) -> Vec<f64>;
     fn motor_count(&self) -> usize;
 
     /// Extracts one field by header string name.
     /// Convenience for system boundaries (WASM bridge, CLI user input).
-    fn field_by_name(&self, name: &str) -> Vec<i64> {
+    fn field_by_name(&self, name: &str) -> Vec<f64> {
         self.field(&SensorField::from_header(name))
     }
 }
